chore(common/compiled): separate type definitions into per-module types.ts files
