+++ common/compiled/node/comms/telegram2/inbound.ts
@@ -1,118 +1,4 @@
-// ─── Telegram API raw object shapes (subset used by inbound parsing) ─────────
-
-interface TgUser {
-  id: number;
-  is_bot: boolean;
-  first_name: string;
-  last_name?: string;
-  username?: string;
-  language_code?: string;
-}
-
-interface TgChat {
-  id: number;
-  type: 'private' | 'group' | 'supergroup' | 'channel';
-  title?: string;
-  username?: string;
-  first_name?: string;
-  last_name?: string;
-}
-
-interface TgEntity {
-  type: string;
-  offset: number;
-  length: number;
-  url?: string;
-  user?: TgUser;
-  language?: string;
-}
-
-interface TgMessage {
-  message_id: number;
-  date: number;
-  edit_date?: number;
-  from?: TgUser;
-  chat: TgChat;
-  text?: string;
-  caption?: string;
-  photo?: Array<{ file_id: string; file_unique_id: string; width: number; height: number; file_size?: number }>;
-  video?: {
-    file_id: string;
-    file_unique_id: string;
-    width: number;
-    height: number;
-    duration: number;
-    mime_type?: string;
-    file_size?: number;
-  };
-  audio?: {
-    file_id: string;
-    file_unique_id: string;
-    duration: number;
-    performer?: string;
-    title?: string;
-    mime_type?: string;
-    file_size?: number;
-  };
-  voice?: { file_id: string; file_unique_id: string; duration: number; mime_type?: string; file_size?: number };
-  document?: { file_id: string; file_unique_id: string; file_name?: string; mime_type?: string; file_size?: number };
-  sticker?: {
-    file_id: string;
-    file_unique_id: string;
-    width: number;
-    height: number;
-    is_animated: boolean;
-    is_video: boolean;
-    emoji?: string;
-    set_name?: string;
-  };
-  location?: { latitude: number; longitude: number; horizontal_accuracy?: number; live_period?: number };
-  contact?: { phone_number: string; first_name: string; last_name?: string; user_id?: number };
-  poll?: {
-    id: string;
-    question: string;
-    options: Array<{ text: string; voter_count: number }>;
-    total_voter_count: number;
-    is_closed: boolean;
-    is_anonymous: boolean;
-    type: string;
-    correct_option_id?: number;
-  };
-  video_note?: { file_id: string; file_unique_id: string; length: number; duration: number; file_size?: number };
-  entities?: TgEntity[];
-  caption_entities?: TgEntity[];
-  reply_to_message?: TgMessage;
-  forward_date?: number;
-  forward_from?: TgUser;
-  forward_from_chat?: TgChat;
-  forward_from_message_id?: number;
-  forward_sender_name?: string;
-  media_group_id?: string;
-}
-
-interface TgUpdate {
-  update_id: number;
-  message?: TgMessage;
-  edited_message?: TgMessage;
-  channel_post?: TgMessage;
-  edited_channel_post?: TgMessage;
-  inline_query?: { id: string; from: TgUser; query: string; offset: string };
-  callback_query?: { id: string; from: TgUser; message?: TgMessage; data?: string; game_short_name?: string };
-  my_chat_member?: {
-    chat: TgChat;
-    from: TgUser;
-    date: number;
-    old_chat_member: { status: string };
-    new_chat_member: { status: string };
-  };
-  chat_member?: {
-    chat: TgChat;
-    from: TgUser;
-    date: number;
-    old_chat_member: { status: string };
-    new_chat_member: { status: string };
-  };
-}
+import type { TgMessage, TgUpdate } from './types.ts';
 
 // ─── Extractors ───────────────────────────────────────────────────────────────
