+++ common/compiled/node/errors/types.ts
@@ -1,0 +1,8 @@
+export interface NormalizedError {
+  statusCode: number;
+  code: string;
+  message: string;
+  details?: unknown;
+  stack?: string;
+  isOperational: boolean;
+}
