+++ common/compiled/node/t4t/custom/index.ts
@@ -1,7 +1,7 @@
 import { parse as csvParse } from 'csv-parse';
 import type { Response } from 'express';
-import * as svc from '../../../services/index.ts';
-import type { T4TRequest } from '../t4t-utils.ts';
+import * as svc from '../../services/index.ts';
+import type { T4TRequest } from '../types.ts';
 
 // custom function demo
 // when country table is uploaded, state data is also created (csv needs to have sufficient data for state row)
