+++ common/compiled/node/ws/types.ts
@@ -1,0 +1,5 @@
+import type WebSocket from 'ws';
+
+export interface AliveWebSocket extends WebSocket {
+  isAlive: boolean;
+}
